+++ NEWS.md
@@ -1,2 +1,8 @@
+# hmma 1.1.0 #
+* Improved the *visualise* function, it now has an extra parameter to avoid 
+  drawing low probability arcs that clutter the drawn graph. Furthermore, the
+  width of an edge is per default related to the weight of an edge. This relation
+  can be linear of via a sigmoid. The minimum and maximum width can be specified.
+  
 # hmma 1.0.0
 First release, initial version.